Add files via upload
This shows the Field Service Project conducted to optimise the firm's existing marketing strategies with the objective of helping the firm advance into the next stage of expansion and growth. R programming and Excel data analytical tools are mainly used for data analysis and visualization and applicable business strategies are used for overall problem solving.